Fix toHaveStyle matcher
The matcher no longer depend on 3rd party libraries to parse and
validate the style values of elements. Instead, it expects the assertion
value to match what getComputedStyle API would return. Another side-effect
of this is that it no longer accepts string values as assertables. This makes
the whole API more uniform and at the same time, less complex to maintain.